refactor: move same-net trace merging to a new pipeline phase
